Fixed player sprite layer
Script attached to player to dynamically change the layering of player sprite at runtime.
Results in player being layered behind walls.
Presents new bug such that when player is in front of a wall and near it they are rendered behind it. Might have to try changing sorting point on player to a custom point. By doing so i hope to have my cake and eat it too.