Attempting to use tile locations to further define sprite layering of player
Ultimate goal, detect what tile player is standing on when colliding with underwall collider.
If player is standing on building floor tile, we know that the player is inside building and when up against the wall should be layered above the wall.
If the player is standing on an environmental tile then the player is outside and should be layered behind the wall.
The particular code i added is not providing desired outcome so i will have to use other functions to figure this out.