Minimized + debugged crafting code.
Minimized crafting code  in order to have crafting table check if it can craft when interacting with crafting slots. I also fixed a bug where a player could craft an item even after removing an ingredient from crafting table slots. Next item is to fix bug of when a player has correctly arranged the ingredients for a craft-able item and instead of the items sprite rendering on the crafting table  a white square is in its place.