Fishing Minigame Oscillation Logic Reframed + Fixed.
Zone of GOod Tension UI component is now a header on a slider so that the headers location with respected to time is clamped to either end of the slider. Future Tooltips and documentation will be needed to fine tine this logic.
Next step implment paremeters form fishing rod and player to influence experience in the fishing mini-game.